# Conflicts:
#	routes/web.php

+++ app/Http/Controllers/RoleController.php
@@ -112,7 +112,7 @@ class RoleController extends Controller
 
         $role->syncPermissions($request->permission_id);
 
-        return redirect()->back()->with('success', 'El recurso se ha actualizado correctamente.');
+        return back()->with('success', 'El recurso se ha actualizado correctamente.');
     }
 
     /**
@@ -124,6 +124,10 @@ class RoleController extends Controller
     public function destroy(Role $role)
     {
         $this->authorize('delete', [Role::class, $role]);
+
+        if (in_array($role->id, [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21])) {
+            return back()->with('error', 'El recurso no se puede eliminar.');
+        }
 
         $role->delete();
